Creating custom hook for page title
