+++ modules/admin-top-bar/assets/js/hooks/use-page-title/use-page-title.js
@@ -1,0 +1,17 @@
+import { useState, useEffect } from 'react';
+
+export const usePageTitle = () => {
+	const [ pageTitle, setPageTitle ] = useState( 'Elementor' );
+
+	useEffect( () => {
+		const pageTitleElement = document.querySelector( '.wp-heading-inline' );
+		if ( ! pageTitleElement ) {
+			return;
+		}
+		setPageTitle( pageTitleElement.innerText );
+
+		pageTitleElement.remove();
+	}, [ ] );
+
+	return pageTitle;
+};
